+++ python/src/scippneutron/load_nexus.py
@@ -3,49 +3,13 @@
 # @author Matthew Jones
 
 import scipp as sc
+from ._loading_common import find_by_nx_class
 from ._loading_detector_data import load_detector_data
 from ._loading_log_data import load_logs
-
 import h5py
 from timeit import default_timer as timer
-from typing import Union, Tuple, Dict, List
+from typing import Union
 from contextlib import contextmanager
-
-
-def _get_attr_as_str(h5_object, attribute_name: str):
-    try:
-        return h5_object.attrs[attribute_name].decode("utf8")
-    except AttributeError:
-        return h5_object.attrs[attribute_name]
-
-
-def _find_by_nx_class(
-        nx_class_names: Tuple[str, ...],
-        root: Union[h5py.File, h5py.Group]) -> Dict[str, List[h5py.Group]]:
-    """
-    Finds groups with requested NX_class in the subtree of root
-
-    Returns a dictionary with NX_class name as the key and list of matching
-    groups as the value
-    """
-    groups_with_requested_nx_class: Dict[str, List[h5py.Group]] = {
-        class_name: []
-        for class_name in nx_class_names
-    }
-
-    def _match_nx_class(_, h5_object):
-        if isinstance(h5_object, h5py.Group):
-            try:
-                nx_class = _get_attr_as_str(h5_object, "NX_class")
-                if nx_class in nx_class_names:
-                    groups_with_requested_nx_class[nx_class].append(h5_object)
-            except KeyError:
-                pass
-
-    root.visititems(_match_nx_class)
-    # Also check if root itself is an NX_class
-    _match_nx_class(None, root)
-    return groups_with_requested_nx_class
 
 
 @contextmanager
@@ -102,7 +66,7 @@ def load_nexus(data_file: Union[str, h5py.File], root: str = "/", quiet=True):
         nx_log = "NXlog"
         nx_entry = "NXentry"
         nx_instrument = "NXinstrument"
-        groups = _find_by_nx_class(
+        groups = find_by_nx_class(
             (nx_event_data, nx_log, nx_entry, nx_instrument), nexus_file[root])
 
         if len(groups[nx_entry]) > 1:
